+++ pyproject.toml
@@ -61,13 +61,17 @@ optional-dependencies.tests = [
   "markdown2",
   "olefile",
   "packaging",
-  "pyarrow",
   "pyroma",
   "pytest",
   "pytest-cov",
   "pytest-timeout",
   "trove-classifiers>=2024.10.12",
 ]
+
+optional-dependencies.test-arrow = [
+  "pyarrow",
+]
+
 optional-dependencies.typing = [
   "typing-extensions; python_version<'3.10'",
 ]
